requirements.txt
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
typer>=0.9.0
streamlit>=1.28.0
matplotlib>=3.7.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
plotly>=5.15.0
numpy>=1.24.0
scipy>=1.10.0
